convert concern into a service object
