+++ app/controllers/sign_up/registrations_controller.rb
@@ -2,7 +2,6 @@ module SignUp
   class RegistrationsController < ApplicationController
     include PhoneConfirmation
     include ApplicationHelper # for ial2_requested?
-    include VendorOutageConcern
 
     before_action :confirm_two_factor_authenticated, only: [:destroy_confirm]
     before_action :require_no_authentication
@@ -71,7 +70,11 @@ module SignUp
     def redirect_if_ial2_and_vendor_outage
       return unless ial2_requested?
 
-      redirect_if_outage(vendors: IAL2_VENDORS, from: CREATE_ACCOUNT)
+      vendor_status = VendorStatus.new
+      if vendor_status.any_ial2_vendor_outage?
+        session[:vendor_outage_redirect] = CREATE_ACCOUNT
+        return redirect_to vendor_outage_url
+      end
     end
   end
 end
