+++ app/controllers/vendor_outage_controller.rb
@@ -1,16 +1,10 @@
 class VendorOutageController < ApplicationController
-  include VendorOutageConcern
-
   def show
-    @specific_message = outage_message
-    tracking_data = {
-      vendor_status: {
-        acuant: IdentityConfig.store.vendor_status_acuant,
-        lexisnexis_instant_verify: IdentityConfig.store.vendor_status_lexisnexis_instant_verify,
-        lexisnexis_trueid: IdentityConfig.store.vendor_status_lexisnexis_trueid,
-      },
-      redirect_from: session.delete(:vendor_outage_redirect),
-    }
-    analytics.track_event(Analytics::VENDOR_OUTAGE, tracking_data)
+    vendor_status = VendorStatus.new(
+      sp: current_sp,
+      from: session.delete(:vendor_outage_redirect),
+    )
+    @specific_message = vendor_status.outage_message
+    vendor_status.track_event(analytics)
   end
 end
